android: Update target SDK to 30
Because Android Studio is being annoying about it, *strongly* suggesting
you target the minimum allowed SDK on Google Play. Coincidentally, this
is also the version of the SDK that my phone uses, so I let it pass.

Note that *minimum* SDK level is still set at 21, we should still be
able to run on older devices. Hopefully someone tests that, lol.