android: Update Kotlin and other dependencies
Because why not, I'm on a roll. Clicking clickety things in the IDE.
These are only used in tests.

Unfortunately, Kotlin cannot be updated past 1.4.x branch (current
one is 1.7.x) because it desugars into some too new Java 1.8 stuff
which is not supported on Android API 21. We'd have to bump our
minimum API level to use newer versions. So not today.