android: CMake build for libthemis_jni
Here's a thing. New SDK and NDK build native code differently because of
course they do. For example, no more .externalNativeBuild directory
with a single build, instead there is .cxx directory with separate
debug and release builds under pseudorandom subdirectory.

Good news is that Google has finally implemented a way to share native
libraries via AARs. Right now we're building BoringSSL as a separate
project and it technically produced an AAR before. But it could not
be reused by other AARs, we directed the build to look at headers
in our source tree and picked up binaries from the build tree.

The new way -- Prefab -- adds a whole bunch of complexity that abstracts
out native library building process, introduces a bunch of structure
into how native libraries are distributed, ultimately allowing to pack
your native libraries into reusable AAR artifacts.

Bad news is that this is Google project, so obviously it was implemented
by someone aiming for a promotion, adds a whole bunch of complexity to
solve 80% of the problem -- except that *our* project is in that 20%
which does not fucking work out of the fucking box and there is not way
to fix it because of all abstraction.

Main problem here is that BoringSSL uses CMake build system while
libthemis_jni is still built using ndk-build, and that combo is not
very well-supported withing the same project. There are reports that
Prefab actually works in the consumer project is build using CMake.

However.

If I were to implemement CMake build for libthemis_jni, then I can
simply add_subdirectory() with BoringSSL and build it right there.

As long as we don't actually planning to distribute our builds of
BoringSSL separately -- and we probably don't, judging how well that
worked out for OpenSSL on iOS -- as long as we don't need AARs,
we don't really need Prefab.

So that's what I'm gonna do. Write a CMake build script for
libthemis_jni and use it for all native code of Android Themis.