android: Update Gradle, Android plugin, build tools, NDK (#942)
* android: Update Gradle, Android plugin, build tools, NDK

Bumping Android Gradle plugin to 7.2.2 so that it picks up new SDKs
and bumping Gradle to 7.3.3 so that it works with the new plugin,
and Android build tools to 30.0.3 so that Gradle works, and with
those new tools NDK has to be configured differently (finally,
it can be pinned in the project configuration itself, instead of
being a free variable in the build machine configuration).

This is quite a YOLO upgrade over multiple major versions, but I don't
think we have a choice. The treadmill must roll forward, because how
else we can Innovate™ to keep the number going up. Otherwise, I can't
build Android Themis anymore with today's tooling.

* android: Update target SDK to 30

Because Android Studio is being annoying about it, *strongly* suggesting
you target the minimum allowed SDK on Google Play. Coincidentally, this
is also the version of the SDK that my phone uses, so I let it pass.

Note that *minimum* SDK level is still set at 21, we should still be
able to run on older devices. Hopefully someone tests that, lol.

* android: Remove JCenter repositories

New Android stuff is finally available in Google repos and we don't
need JCenter anymore. Good night, sweet prince.

* android: Update Kotlin and other dependencies

Because why not, I'm on a roll. Clicking clickety things in the IDE.
These are only used in tests.

Unfortunately, Kotlin cannot be updated past 1.4.x branch (current
one is 1.7.x) because it desugars into some too new Java 1.8 stuff
which is not supported on Android API 21. We'd have to bump our
minimum API level to use newer versions. So not today.

* android: New way to filter out annotations

Android has its own annotations under "android" and they conflict
with Jetbrains annotations. I was never motivated enough to unify
them so instead we've been filtering them out for Android builds.

New Gradle -- new approach. Such innovative. Much simpler.

* android: CMake build for libthemis_jni

Here's a thing. New SDK and NDK build native code differently because of
course they do. For example, no more .externalNativeBuild directory
with a single build, instead there is .cxx directory with separate
debug and release builds under pseudorandom subdirectory.

Good news is that Google has finally implemented a way to share native
libraries via AARs. Right now we're building BoringSSL as a separate
project and it technically produced an AAR before. But it could not
be reused by other AARs, we directed the build to look at headers
in our source tree and picked up binaries from the build tree.

The new way -- Prefab -- adds a whole bunch of complexity that abstracts
out native library building process, introduces a bunch of structure
into how native libraries are distributed, ultimately allowing to pack
your native libraries into reusable AAR artifacts.

Bad news is that this is Google project, so obviously it was implemented
by someone aiming for a promotion, adds a whole bunch of complexity to
solve 80% of the problem -- except that *our* project is in that 20%
which does not fucking work out of the fucking box and there is not way
to fix it because of all abstraction.

Main problem here is that BoringSSL uses CMake build system while
libthemis_jni is still built using ndk-build, and that combo is not
very well-supported withing the same project. There are reports that
Prefab actually works in the consumer project is build using CMake.

However.

If I were to implemement CMake build for libthemis_jni, then I can
simply add_subdirectory() with BoringSSL and build it right there.

As long as we don't actually planning to distribute our builds of
BoringSSL separately -- and we probably don't, judging how well that
worked out for OpenSSL on iOS -- as long as we don't need AARs,
we don't really need Prefab.

So that's what I'm gonna do. Write a CMake build script for
libthemis_jni and use it for all native code of Android Themis.

* android: Use CMake instead of ndk-build

Now that the script is there, replace ndk-build with CMake.

While we're here, pin the verison of CMake in build.gradle
so that it's more reproducible.

* android: Remove ndk-build Makefiles

I don't think we'd ever want to go back to ndk-build, drop it.

* android: Remove ":boringssl" project

Since now BoringSSL is a part of ":android" native code, there is no
need to build it as a separate project. We don't have other users.

* android: Roll back to NDK 25.0.8775105

Harmonize NDK version with what's preinstalled in GitHub Actions
environments, shaving off some minutes from build time. We don't
particularly care about the NDK version, as long as it's some
modern version that might be supported for some next years.

* android-ci: Switch to macos-11

It has newer Android build tools. And yes, still macOS, because running
Android emulator on Linux still sucks on GitHub Actions.

* android-ci: Remove old CMake hack

No longer necessary. Newer Android tooling is able to filter that out,
and it's not installed in newer environments anyway.

* android-ci: Use Java 11 for Gradle builds

But Java 8 for Android's sdkmanager. IKR

* android-ci: Install SDK & NDK explicitly

Just in case CI environment changes. They typically don't change
defaults and keep stuff installed, but who knows. Be explicit.

* android: Update publishing artifact name

New SDK calls it "android-release.aar", not just "android.aar".