themis: THEMIS_GEN_EC_KEY_PAIR_UNCOMPRESSED switch
Generate uncompressed public keys when THEMIS_GEN_EC_KEY_PAIR_UNCOMPRESSED
environment variable is set to 1. This is an opt-in for arguable better
representation of public keys.

Add a smoke test for this functionality so that "make test" confirms
that it works. The next steps will run the entire test suite with that
variable.