pythemis: Correct Metadata-Version
I don't know how was the first to have an idea that this field is
for the package version, instead of being the version of PKG-INFO
metadata itself. But this field has been updated like clockwork.

I doesn't seem that anybody cared, not Python at least. Set this version
to the latest 1.2.