CI: Test Themis Core on Ubuntu 22.04
So here's the story. Themis does not (yet) support OpenSSL 3.0. There
is an experimental flag to enable support, but we haven't tested it all
too well. It probably works, but we aren't ready to sign off on that.
Thus the build fails with OpenSSL 3.0 by default.

Here's the situation with CI:

  - Ubuntu 20.04 has only OpenSSL 1.1.1 (which is the default)

  - Ubuntu 22.04 has only OpenSSL 3.0 (which is the default)

  - macOS can have both, using OpenSSL 1.1.1 if available,
    or failing the build otherwise.

Make sure to test all these expectations. Note that most of the jobs
are still running on Ubuntu 20.04 now. Ubuntu 22.04 is tested with
OpenSSL 3.0 only for the core library.