rust: Suppress "clippy::needless_borrow"
Latest Clippy got this new lint and it's, like, everywhere in the
codebase. The damned stationery is on a crusage against undesireable
ampersands. All because RustThemis codebase likes to use AsRef<[u8]>
instead of just &[u8], so now passing [] is strictly preferable to &[]
as well as passing `some_vector` is preferable to `&some_vector`
if the vector is not used anywhere else (as it often is in tests).

Initially I wanted to oblige the piece of metal, but after fixing up
wave after wave of warnings I decided that it is enough and I won't be
a slave to the machine, given that in some cases the changes make the
code less aesthetically pleasing.