rust: Suppress "clippy::needless_borrow" (#971)
* rust: Suppress "clippy::needless_borrow"

Latest Clippy got this new lint and it's, like, everywhere in the
codebase. The damned stationery is on a crusage against undesireable
ampersands. All because RustThemis codebase likes to use AsRef<[u8]>
instead of just &[u8], so now passing [] is strictly preferable to &[]
as well as passing `some_vector` is preferable to `&some_vector`
if the vector is not used anywhere else (as it often is in tests).

Initially I wanted to oblige the piece of metal, but after fixing up
wave after wave of warnings I decided that it is enough and I won't be
a slave to the machine, given that in some cases the changes make the
code less aesthetically pleasing.

* CI: Test RustThemis with Rust 1.51 and stable

Rust 1.51 is the minimal version that we support, currently limited
by "zeroize" dependency. Make sure it stays that way.

* CI: Rename "whitelist" => "allowlist"

Here, I hope this is less offensive for you, ffs.

I'm triggered by rustc using "-Z" for its unstable flags, but I don't
raise issues to rename it into something different.