CI: Pin Rust toolchain at 1.56
The MSRV stated by the changelog is 1.47 and initially I wanted to be
testing with that version. However, our dev-dependencies don't consider
MSRV changes in their dependencies to be a breaking change, and thus
it's not possible to build RustThemis tests with Rust 1.47 since some
transitive dependencies require Rust 2021 edition available since 1.56.

In theory, I could go and restrict the maximum versions of dependencies
to keep things working on 1.47. But that's too much effort for something
that nobody should really need.

But! I don't want to change any code either unless necessary. Using 1.56
does not require any changes, so no changes will be made.

Keep testing with 1.56 but since this is a maintenance branch, don't
test with the latest stable. It won't work.