rust: Use f-string formatting syntax
Or "variable capture" as it's called around here. Since of Rust 1.67,
our God-Emperor Clippy started demanding this form of formatting.

    error: variables can be used directly in the `format!` string
      --> src/wrappers/themis/rust/libthemis-sys/build.rs:49:13
       |
    49 |             eprintln!("{}", error);
       |             ^^^^^^^^^^^^^^^^^^^^^^
       |
       = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#uninlined_format_args
       = note: `-D clippy::uninlined-format-args` implied by `-D warnings`
    help: change this to
       |
    49 -             eprintln!("{}", error);
    49 +             eprintln!("{error}");
       |

Given that we're running WITH_FATAL_WARNINGS, we must either cave in,
or suppress the lint. Suppressing it on Rust 1.56 requires suppressing
another lint about unknown lint being suppressed.

So, given that we're unlikely to have users which require Rust 1.56
specifically, I think it's okay to cave in, update MSRV to 1.58, and
update all affect format strings so that Clippy stays happy.