jsthemis: Add missing returns after errors
There were two places in the code with missing return statements
after errors. As a result, if something in key generation goes
wrong, the wrapper will segfault while trying to allocate huge
vectors for the keys.

It's interesting how these bugs remained unnoticed for such a
long time (~4 years). This is because it's hard to make key pair
generation fail. It works surprisingly well most of the time :)