jsthemis: Downgrade mocha to ^7
This version and all its transitive dependencies support node v8
which is the only version we can run on centos 7. It seems like other
OS work fine, at least on buildbot.

Another option is to fix our scripts on buildbot just for centos 7 so
they patch the versions before testing. This may save us from future
problems. However, downgrading dev dependency is not that scary, so
maybe this approach is okay.