rust-themis: Run the skipped tests on the 32bit
They were skipped because allocation on 32-bit platforms failed
while trying to allocate something bigger than 2GB. This is due to
usage of `.reserve` which panics if it couldn't fulfill the request.

Since we traversed to the rust 1.58, we now have the `try_*` methods,
including the `.try_reserve` which allows us handle panics gracefully.

I've tested it manually on pi4 and it works!