make/install libthemis deb in Python CI job
Since python3-pythemis now depends on libthemis, we need to install the
dependency before installing Python wrapper.

Updated Makefile to conditionally include C++ and Java wrappers in deb
target so that one could build only Themis itself on systems without
Java (jni.h is required to to build Java wrapper). This may not be the
best way, but it uses the least amount of changes.

Potential improvements:
* Split deb target into deb_themis, deb_themispp, deb_java (will need to
  copy-paste some vars as well, and to the same to rpm for consistency)
* Add new target like deb_themis_install and others, just like did it
  with pythemis_install_deb